feat(ui): style moved to modifier
